requirements.txt
fastapi[standard] >=0.115.12
uvicorn[standard] >= 0.30.0
alembic>= 1.16.1
pytest >= 8.3.5
uvicorn[standard]>=0.34.2
SQLAlchemy>= 2.0.41
psycopg2-binary>=2.9.10
bcrypt>=4.3.0
python-jose>=3.5.0
passlib>=1.7.4